Add files via upload
The loan prediction project in machine learning involves cleaning the dataset and training multiple models, such as Logistic Regression, Random Forest, SVM, Decision Tree, and Gradient Boosting. After training, the models are evaluated using a testing dataset to calculate their accuracy. The model with the highest accuracy is considered the most accurate for predicting loan approvals. The results are analyzed to determine the best model for the project.